+++ fpgasim/frontend/src/components/GridRenderer.tsx
@@ -1,0 +1,21 @@
+import * as d3 from "d3";
+
+export function renderGrid(
+  gridGroup: d3.Selection<SVGGElement, unknown, null, undefined>,
+  cols: number,
+  rows: number,
+  cellSize: number
+) {
+  for (let i = 0; i < cols; i++) {
+    for (let j = 0; j < rows; j++) {
+      gridGroup
+        .append("rect")
+        .attr("x", i * cellSize)
+        .attr("y", j * cellSize)
+        .attr("width", cellSize)
+        .attr("height", cellSize)
+        .attr("stroke", "black")
+        .attr("fill", "none");
+    }
+  }
+}
