+++ fpgasim/frontend/src/components/NodeRenderer.tsx
@@ -1,0 +1,59 @@
+import * as d3 from "d3";
+import { getColor, getStrokeColor } from "../utils/colorsUtils";
+import { getTooltip } from "../utils/toolTipsUtils";
+import { TimingCell } from "../types";
+
+// If you rely on `.index`, extend the base TimingCell type
+type ExtendedTimingCell = TimingCell & { index?: number };
+
+export function renderNodes(
+  nodeGroup: d3.Selection<SVGGElement, unknown, null, undefined>,
+  cells: ExtendedTimingCell[],
+  cellSize: number,
+  cols: number,
+  currentStep: number,
+  isPlaying: boolean
+) {
+  cells.forEach((cell, index) => {
+    const match = cell.instance.match(/_(\d+)_(\d+)(?!.*\d)/);
+    const xCoord = match ? parseInt(match[1], 10) : index % cols;
+    const yCoord = match ? parseInt(match[2], 10) : Math.floor(index / cols);
+
+    const x = xCoord * cellSize;
+    const y = yCoord * cellSize;
+    const isActive = index === currentStep;
+
+    const group = nodeGroup
+      .append("g")
+      .attr("transform", `translate(${x + cellSize / 2}, ${y + cellSize / 2})`);
+
+    const circle = group
+      .append("circle")
+      .attr("r", cellSize / 3.5)
+      .attr("fill", getColor(cell.cell_type))
+      .attr("stroke", isActive ? "#FFD700" : getStrokeColor(cell.cell_type))
+      .attr("stroke-width", isActive ? 4 : 1.5);
+
+    if (!isPlaying) {
+      circle
+        .on("mouseover", function () {
+          d3.select(this).attr("stroke-width", 3);
+        })
+        .on("mouseout", function () {
+          d3.select(this).attr("stroke-width", 1.5);
+        });
+    }
+
+    group.append("title").text(getTooltip(cell));
+
+    if (isActive) {
+      group
+        .append("text")
+        .text("Active")
+        .attr("text-anchor", "middle")
+        .attr("dy", cellSize / 2 + 12)
+        .style("font-size", "10px")
+        .style("fill", "#333");
+    }
+  });
+}
