+++ fpgasim/frontend/src/utils/sharedStyles.ts
@@ -1,0 +1,10 @@
+export const buttonStyle = {
+    padding: "6px 12px",
+    backgroundColor: "#1976d2",
+    color: "#fff",
+    border: "none",
+    borderRadius: "4px",
+    cursor: "pointer",
+    fontSize: "13px",
+  };
+
